% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn.R
\docType{data}
\name{burn}
\alias{burn}
\title{STEP file S0100120.burn}
\format{
A tibble
}
\usage{
burn
}
\description{
burn : is an example of STEP file ".burn" used to enter the proportion of burned area ([0-1]).
Even if no fire took place during the year, the file must be created
}
\keyword{datasets}
