% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha.R
\docType{data}
\name{cha}
\alias{cha}
\title{Example of STEP file S0100120.cha}
\format{
A tibble. Example of a STEP file S0100120.cha: an input file of STEP
}
\usage{
cha
}
\description{
cha : is an example of .cha file used for animal load
Structure: 1 Line: type of distribution of animals by category (Cattle) (Goats) (Sheep) (Donkeys) (Camelins) (Horse)
           31 Lines: Day, total animal load (number UBT/ha) * 12
}
\keyword{datasets}
