% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cle.R
\docType{data}
\name{cle}
\alias{cle}
\title{STEP file S0100120.cle}
\format{
A tibble
}
\usage{
cle
}
\description{
cle : is an example of STEP file cle used for STEP calibration parameters
}
\details{
\describe{
  \item{1 Line:}{ conversion efficiency (gMS/MJ)}
  \item{1}{initial green biomass (gMS/m2)}
  \item{2}{Percentage of C3 (\%)}
  \item{3}{Initial slag (cm² g-1)}
}
}
\keyword{datasets}
