% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_out_dir.R
\name{clean_out_dir}
\alias{clean_out_dir}
\title{Back up outputs of a previous simulation}
\usage{
clean_out_dir(new_outpath, old_outpath)
}
\arguments{
\item{new_outpath}{the path to the folder "new_output" (the current directory of the outputs)}

\item{old_outpath}{the path to a back up folder "old_output" (the folder in which you want to store the outputs of the previous simulation)}
}
\value{

}
\description{
Back up outputs of a previous simulation
}
\examples{
\dontrun{
new_outpath = normalizePath("2-outputs/new_output/", winslash = "/")
old_outpath = normalizePath("2-outputs/old_output/", winslash = "/")

clean_out_dir(new_outpath,old_outpath)
}


}
