% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFlux.R
\name{CO2equi}
\alias{CO2equi}
\title{Convert flux data to CO2 equivalent}
\usage{
CO2equi(x, Gas)
}
\arguments{
\item{x}{flux data in CO2-C or CH4-C or N2O-N}

\item{Gas}{unit of the gas CO2-C or CH4-C or N2O-N}
}
\value{
flux value in CO2 equivalent
}
\description{
CH4 and N2O emissions rates can be converted to CO2 equivalents using the ‘100-yr global warming potential’ factors as follows:
1 kg N2O = 298 kg CO2 equivalents. 1 kg CH4 = 25 kg CO2 equivalents.
Please note that, according the most recent IPCC report (IPCC, 2013),
the conversion of CH4 into CO2eq. should be:  1 kg CH4 = 34 kg CO2 equivalents
Reference cited: IPCC (2013) Climate change 2013: the physical science basis. Intergovernmental Panel on Climate Change. Available at https://www.ipcc.ch/report/ar5/wg1/
"1 kg CH4-C = (16/12)x1 kg CH4 = 1.33 kg CH4". "1 kg CO2-C = (44/12)x1 kg CO2 =3.67 kg CO2". "1 kg N2O-N = (44/28)*1 kg N2O = 1.57 kg N2O"
}
\examples{
\dontrun{
CO2equi(0.6,Gas="N2O-N")
}

}
