% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{STEP file pluri.dat}
\format{
A tibble
}
\usage{
dat
}
\description{
dat : is an example of file generated by interface to allow multi-year execution of STEP
In this example we have 9 years of simulation where 2012 is repeated 6 times (6 years of spine-up)
}
\details{
\describe{
  \item{1 Line:}{number of years chosen "N". For a simple annual run, N=1}
  \item{N Lines:}{8-character prefixes "S0100112". Uniquely designates a given year, for a given country, for a given site, and for a given situation}
  \item{In}{S0100120.cha}
  \item{S:}{the initial of the country, here Senegal}
  \item{01:}{the number of the site}
  \item{001:}{the experimental situation}
  \item{12:}{the year of the simulation, here 2012}
}
}
\keyword{datasets}
