% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\docType{data}
\name{disp}
\alias{disp}
\title{STEP file S0100120.disp}
\format{
A tibble
}
\usage{
disp
}
\description{
disp: is an example of STEP file ".dis" used to provide information about the mode of dispersal of diaspores
}
\details{
\describe{
  \item{1 Line:}{proportions of the types of dispersion}
  \item{1:}{wet season}
  \item{2:}{very early}
  \item{3:}{early}
  \item{4:}{progressive}
  \item{5:}{late}
  \item{6:}{very late}
  \item{}{if the first value = -1 (default values)}
}
}
\keyword{datasets}
