% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_step_var.R
\name{find_step_var}
\alias{find_step_var}
\title{Information about STEP outputs}
\usage{
find_step_var(keyword)
}
\arguments{
\item{keyword}{A keyword of the sought output}
}
\value{
The name of the output sought and its meaning
}
\description{
Find step output names and their meanings
}
\examples{
\dontrun{
find_step_var("Shum")
}

}
