% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_file.R
\name{gen_step_file}
\alias{gen_step_file}
\title{Write STEP2D input files for each site}
\usage{
gen_step_file(filepath, filename, state.date, end.date, isite)
}
\arguments{
\item{filepath}{path to where you want to store the file generated}

\item{filename}{the name of the initialisation file you want to generate (the extension)}

\item{state.date}{the two last digits of the first year of simulation. e.g. for 2012 it is "12"}

\item{end.date}{the two last digits of the last year of simulation.. e.g. for 2020 it is "20"}

\item{isite}{id of the site. e.g. "S02", usually the name of the folder in which you want to store the data}
}
\value{
STEP input file formats : ".dat",".bmt",".burn",".cha",".cle",".disp",".sol",".veg",".wet"
}
\description{
gen_step_file generates examples of STEP initialisation file in excel format
}
\examples{
\dontrun{
gen_step_file(filepath = filepath,filename ="cha",
}

}
