% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_mto.R
\name{gen_step_mto}
\alias{gen_step_mto}
\title{Generate meteo files for STEP model}
\usage{
gen_step_mto(workspace, dataframe, isite, year, alt, lat, hautmes)
}
\arguments{
\item{workspace}{where you want to save the mto file}

\item{dataframe}{the name of the dataframe to convert to mto file}

\item{isite}{the name of the folder in which the mto file will be stored.
Usually the name if the site.}

\item{year}{the two last digits of the year. e.g. for 2012 it is "12"}

\item{alt}{the altitude of the site}

\item{lat}{the latitude of the site}

\item{hautmes}{the height at which the wind speed was measured}
}
\value{
a folder in the workspace in which there is a mto file. e.g. for
site 01 and year 2012 the file name is S0100112.mto
}
\description{
Formatting csv data in STEP mto format.
The function takes dataframe with: names(dataframe)="date" "day" "year" "rain" "RayGlo_MJm" "temp_min"
 "temp_max" "pvap_2_m_hPa" "Wind_Speed".
}
\examples{
\dontrun{
gen_step_mto(workspace = workspace,dataframe=df,isite="Dahra",
year = "12",alt=308,lat=15.4,hautmes=2)
}

}
