% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_out.R
\name{gen_step_out}
\alias{gen_step_out}
\title{STEP Out files to Excel and CSV}
\usage{
gen_step_out(file.name, workspace, state.date, end.date, format = "csv")
}
\arguments{
\item{file.name}{the name of the file. Usually the name of the site}

\item{workspace}{the directory, where there are the files you want to format.
And where you to store the new files}

\item{state.date}{the two last digits of the first year. e.g. for 2012 it is "12"}

\item{end.date}{the two last digits of the last year of simulation.. e.g. for 2020 it is "20"}

\item{format}{the desired format for the file e.g. xlsx or csv}
}
\value{
a csv or excel file
}
\description{
gen_step_out writes readable output files in the formats xlsx and csv
}
\examples{
\dontrun{
gen_step_out("Dahra",workspace,12,20,"csv")
}


}
