% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_plot.R
\name{gen_step_plot}
\alias{gen_step_plot}
\title{gen_step_plot : plot simulation vs observation}
\usage{
gen_step_plot(
  variable_names,
  df_simu,
  df_obs,
  start_date = NULL,
  end_date = NULL,
  plot_type = "ts"
)
}
\arguments{
\item{variable_names}{the name of your model variables you want to plot}

\item{df_simu}{the name of the dataframe of your model output}

\item{df_obs}{the name of the dataframe of your observations}

\item{start_date}{select a start date if you don't want to plot the whole data and just want to zoom out}

\item{end_date}{select an end date if you don't want to plot the whole data and just want to zoom out}

\item{plot_type}{the type of plot you want. "ts" for time series and "sc" for scatter plot}
}
\value{
time series or scatter plot (with R^2 and RMSE)
}
\description{
gen_step_plot  plot time series and scatter plots of simulation versus observation for selected model variables.
in plot_type choose the type of plot you want. "ts" for time series and "sc" for scatter plot
}
\examples{
\dontrun{
gen_step_plot (variable_names, df_simu, df_obs, start_date=NULL, end_date=NULL,plot_type="ts")
}



}
