% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_soil.R
\name{gen_step_soil}
\alias{gen_step_soil}
\title{Write STEP-GENDEC-N2O soil input file for each simulation grid cell}
\usage{
gen_step_soil(dataframe, start_yr, end_yr, folder_patch)
}
\arguments{
\item{dataframe}{the data frame containing the data required to create the soil file
e.g. of variable names: id, coord, albedo, soil_water1, soil_temp1, clay1, sand1, pH1, silt1,
soil_water2,soil_temp2, clay2, sand2, pH2, silt2,soil_water3, soil_temp3, clay3, sand3, pH3,silt3,
soil_water4, soil_temp4, clay4, sand4, pH4, silt4, soil_temp5}

\item{start_yr}{the two last digits of the first year of simulation. e.g. for 2012 it is "12"}

\item{end_yr}{the two last digits of the last year of simulation.. e.g. for 2020 it is "20"}

\item{folder_patch}{path to where you want to store the file generated}
}
\value{
2D STEP-GENDEC-NC input file formats ".sol" for all simulation grid cells
}
\description{
gen_step_file generates examples of STEP initialisation file in excel format
}
\examples{
\dontrun{
gen_step_soil(dataframe,start_yr,end_yr,folder_patch)
}

}
