% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table.R
\name{gen_table}
\alias{gen_table}
\title{Generate dataframe for processing}
\usage{
gen_table(filepath, workspace, filename, sim_variables, obs = TRUE)
}
\arguments{
\item{filepath}{path of the files obs_data.xlsx and test0_data.xlsx.
Usually the path of the folder daily_original}

\item{workspace}{path of the folder where you stored the STEP
output file you are processing}

\item{filename}{the name of the STEP output file your are processing}

\item{sim_variables}{names of the variables you want to process
e.g. sim_variables = c("Rain","CO2Soil","Shum_1_per")}

\item{obs}{notify whether or not you have observation data of the variables of interest}
}
\value{
a dataframe of the variables of interest. the observation data and the result
of the current and the previous simulation "test0_data", all bind together
}
\description{
gen_table generates dataframe to process selected variables
}
\examples{
\dontrun{
gen_table(filepath = filepath,workspace = workspace,filename = "Dahra",
sim_variables = sim_variables,obs=TRUE)
}



}
