% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_0_1}
\alias{norm_0_1}
\title{Linear Scaling normalization is used when the feature is more-or-less uniformly distributed across a fixed range.
# https://www.indeed.com/career-advice/career-development/normalization-formula}
\usage{
norm_0_1(x)
}
\arguments{
\item{x}{the value in the dataset to normalize between 0 and 1}
}
\value{
the normalized x value
}
\description{
creating a common scale to compare data sets with very different values
}
\details{
Linear Scaling normalization
}
\examples{
\dontrun{
norm_0_1(data$Biomass2019)
}

}
