% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_a_b}
\alias{norm_a_b}
\title{Linear Scaling Normalization formula for custom ranges}
\usage{
norm_a_b(x, a, b)
}
\arguments{
\item{x}{the value in the dataset to normalize between a and b}

\item{a}{the lowest value of the new range}

\item{b}{the highest value of the new range}
}
\value{
the normalized x values
}
\description{
normalize a dataset with in a custom range where the lowest value is a and the highest value is b
}
\examples{
\dontrun{
norm_a_b(data$Biomass2019)
}

}
