% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_1D_step.R
\name{run_1D_step}
\alias{run_1D_step}
\title{Run STEP model from Rstudio}
\usage{
run_1D_step(step_daily_original, daily_original)
}
\arguments{
\item{step_daily_original}{Path to the directory containing the model's code}

\item{daily_original}{Path to the directory containing the model's input files}
}
\value{
Model output files in a folder called Output
}
\description{
Run STEP model using R
}
\examples{
\dontrun{
run_1D_step(step_daily_original,daily_original)
}

}
