% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step.R
\name{run_step_1D}
\alias{run_step_1D}
\title{Run STEP model from R}
\usage{
run_step_1D(
  step_daily_original = step_daily_original,
  daily_original = daily_original
)
}
\arguments{
\item{step_daily_original}{Path to the directory containing the model's code}

\item{daily_original}{Path to the directory containing the model's input files}
}
\value{
Model output files in a folder called Output
}
\description{
Run STEP model using R
}
\examples{
\dontrun{
run_step_1D(step_daily_original,daily_original)
}

}
