% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_plot.R
\name{step_plot}
\alias{step_plot}
\title{Dynamic and scatter plots for selected variables}
\usage{
step_plot(dataframe, varname, type = "dynamic", obs = F)
}
\arguments{
\item{dataframe}{the name of the dataframe generated with the function gen_table}

\item{varname}{the name of the STEP variable}

\item{type}{the type of dynamic plot you want "line or point)}

\item{obs}{notify whether or not you want to display the observations alongside}
}
\value{
a plot "a dynamic time series or a scatter plot"
}
\description{
step_plot displays a dynamic plot or a scatter plot for selected step variable
}
\examples{
\dontrun{
step_plot(dataframe = df, varname = "Etr",type="line",obs=TRUE)
}



}
