% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_reg.R
\name{step_reg}
\alias{step_reg}
\title{Perform a linear regression}
\usage{
step_reg(dataframe, varname, target = NULL)
}
\arguments{
\item{dataframe}{the name of the dataframe generated with the function gen_table}

\item{varname}{the name of the STEP variable}

\item{target}{notify your target, the scatterplot, the metric or the summary list of the regression}
}
\value{
a scatter plot, a summary or metrics table of the regression
}
\description{
step_reg performs a linear regression between observation and simulation.
if target = "plot" it displays the scatter plot of the regression,
if target = "summary" it displays the summary table of the regression,
if target = "metric" it displays the metric of the regression
}
\examples{
\dontrun{
step_reg(dataframe = df,varname = "CO2Soil",target = "metric")
}


}
