% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_var_info.R
\name{step_var_info}
\alias{step_var_info}
\title{Information about STEP outputs}
\usage{
step_var_info(keyword)
}
\arguments{
\item{keyword}{A keyword of the sought output}
}
\value{
The name of the output sought and its meaning
}
\description{
Find step output names and their meanings
}
\examples{
step_var_info("Shum")
\dontrun{
step_var_info("Shum")
}

}
