% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFlux.R
\name{W_to_MJ}
\alias{W_to_MJ}
\title{Global radiation from W to MJ}
\usage{
W_to_MJ(x, unit)
}
\arguments{
\item{x}{global radiation in W}

\item{unit}{time step (e.g. "hour" for MJ/hr, "day", "min", "quarter", "sec", "half")}
}
\value{
global radiation in MJ
}
\description{
W_to_MJ convert radiation from W to MJ
}
\examples{
\dontrun{
W_to_MJ(566,"hour")
}

}
