% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet.R
\docType{data}
\name{wet}
\alias{wet}
\title{STEP file S0100120.wet}
\format{
A tibble
}
\usage{
wet
}
\description{
wet : is an example of STEP file ".wet" used to provide information about flood events
}
\keyword{datasets}
