% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmt.R
\docType{data}
\name{bmt}
\alias{bmt}
\title{Example of STEP file S0100120.bmt}
\format{
A tibble. Example of a STEP file S0100120.bmt: an input file of STEP
}
\usage{
bmt
}
\description{
bmt : is an example of .bmt file used to call optimi.m.
It contains the number of the day of the year corresponding to the value of biom
}
\keyword{datasets}
