% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pvpa_hPa.R
\name{cal_pvap_hPa}
\alias{cal_pvap_hPa}
\title{Compute vapor pressure in hPa}
\usage{
cal_pvap_hPa(relative_humidity, mean_temperature)
}
\arguments{
\item{relative_humidity}{relative humidity in \%}

\item{mean_temperature}{is the mean temperature in deg C}
}
\value{
vapor pressure in hPa
}
\description{
cal_pvap_hPa Computes vapor pressure hPa
}
\examples{
\dontrun{
cal_pvap_hPa(57,20)
}

}
