% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cha.R
\docType{data}
\name{cha}
\alias{cha}
\title{STEP file S0100120.cha}
\format{
A tibble
}
\usage{
cha
}
\description{
cha : is an example of .cha file used for animal load
}
\details{
\describe{
  \item{1 Line:}{ type of distribution of animals by category (Cattle) (Goats) (Sheep) (Donkeys) (Camelins) (Horse)}
  \item{31 Lines:}{Day, total animal load (number UBT/ha) * 12}
}
}
\keyword{datasets}
