% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cle.R
\docType{data}
\name{cle}
\alias{cle}
\title{STEP file S0100120.cle}
\format{
A tibble
}
\usage{
cle
}
\description{
cle : is an example of STEP file ".cle" used for STEP calibration parameters
}
\details{

}
\keyword{datasets}
