% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example of pluri.dat file}
\format{
A tibble. Example of a STEP file Pluri.dat: an input file of STEP
}
\usage{
dat
}
\description{
dat : is an example of file generated by interface to allow multi-year execution of STEP
Structure: 1 Line: number of years chosen "N". For a simple annual run, N=1
           N Lines: 8-character prefixes "S0100112". Uniquely designates a given year, for a given country, for a given site, and for a given situation
           S: Senegal
           01: the number of the site
           001: the experimental situation
           12: the simulation of the year 2012
In this example we have 9 years of simulation where 2012 is repeated 6 times (6 years of spine-up)
}
\keyword{datasets}
