% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_step_xlsx.R
\name{gen_step_xlsx}
\alias{gen_step_xlsx}
\title{Generate an example of STEP initialisation file in excel format}
\usage{
gen_step_xlsx(filepath = NULL, filename)
}
\arguments{
\item{filepath}{path to where you want to store the file generated}

\item{filename}{the name of the initialisation file you want to generate}
}
\value{
An excel file to edit
}
\description{
gen_step_xlsx generates examples of STEP initialisation file in excel format.
filename = "dat" generates pluri.dat file in excel format to edit.
filename = "bmt" generates S100112.bmt file in excel format to edit.
filename = "burn" generates S100112.burn file in excel format to edit.
filename = "cha" generates S100112.cha file in excel format to edit.
filename = "cle" generates S100112.cle file in excel format to edit.
filename = "disp"generates S100112.disp file in excel format to edit.
filename = "sol" generates S100112.sol file in excel format to edit.
filename = "veg" generates S100112.veg file in excel format to edit.
filename = "wet" generates S100112.wet file in excel format to edit.
}
\examples{
\dontrun{
gen_step_xlsx(filepath,"sol")
}


}
